Take full*, Light* out to the mod.rs
