+++ node/service/src/service/mod.rs
@@ -1,8 +1,23 @@
 pub mod crab;
 pub mod darwinia;
 
+pub use darwinia_primitives::OpaqueBlock as Block;
 use sc_service::{config::PrometheusConfig, ChainSpec, Configuration, Error as ServiceError};
+use sp_runtime::traits::BlakeTwo256;
 use substrate_prometheus_endpoint::Registry;
+
+type FullBackend = sc_service::TFullBackend<Block>;
+type FullSelectChain = sc_consensus::LongestChain<FullBackend, Block>;
+type FullClient<RuntimeApi, Executor> = sc_service::TFullClient<Block, RuntimeApi, Executor>;
+type FullGrandpaBlockImport<RuntimeApi, Executor> = sc_finality_grandpa::GrandpaBlockImport<
+	FullBackend,
+	Block,
+	FullClient<RuntimeApi, Executor>,
+	FullSelectChain,
+>;
+type LightBackend = sc_service::TLightBackendWithHash<Block, BlakeTwo256>;
+type LightClient<RuntimeApi, Executor> =
+	sc_service::TLightClientWithBackend<Block, RuntimeApi, Executor, LightBackend>;
 
 /// Can be called for a `Configuration` to check if it is a configuration for the `Crab` network.
 pub trait IdentifyVariant {
